Restrict visibility of libraries that are newly vendor_available: true.
As part of fixing b/68319577, some libraries were made vendor_available: true.
To minimize new dependencies from vendor on these libraries, this change
restricts the visibility of the libraries to only the framework projects that
currently use them.

This change does two things. It makes libbz a static libarary only (drops the
shared library). It restricts the visibility of the static library to only the
framework projects that currently use it.

Bug: 141649008
Test: Targets build successfully.
Change-Id: Icb2974bd890a705f827f98a2526eed21625bf9f4